backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6
pymupdf==1.23.7
sentence-transformers==2.2.2
chromadb==0.4.18
httpx>=0.23.0,<0.24.0
aiofiles==23.2.1
pytest==7.4.3
streamlit==1.32.0
supabase==1.0.3
sseclient-py==1.7.2
reportlab==4.1.0